PartitionedMPSs

Documentation for PartitionedMPSs.

PartitionedMPSs.LazyContraction

PartitionedMPSs.PartitionedMPS

PartitionedMPSs.Projector

PartitionedMPSs.Projector

PartitionedMPSs.SubDomainMPS

Base.:&

Base.:+

Base.:<

Base.:|

Base.getindex

Base.isdisjoint

Base.keys

Base.length

Base.values

LinearAlgebra.norm

PartitionedMPSs._add_patching

PartitionedMPSs._next_projindex

PartitionedMPSs.adaptive_patching

PartitionedMPSs.adaptive_patching

PartitionedMPSs.adaptivecontract

PartitionedMPSs.add_patching

PartitionedMPSs.contract

PartitionedMPSs.extractdiagonal

PartitionedMPSs.makesitediagonal

PartitionedMPSs.projcontract

PartitionedMPSs.projcontract

PartitionedMPSs.project

PartitionedMPSs.rearrange_siteinds

PartitionedMPSs.siteindices

PartitionedMPSs.LazyContraction — Type

Lazy evaluation for contraction of two SubDomainMPS objects.

PartitionedMPSs.PartitionedMPS — Type

PartitionedMPS is a structure that holds multiple MPSs (SubDomainMPS) that are associated with different non-overlapping projectors.

PartitionedMPSs.Projector — Type

A projector represents a projection of a tensor from a set of its indices to integers. Each index is projected to a positive integer.

PartitionedMPSs.Projector — Method

Constructing a projector from a single pair of index and integer.

PartitionedMPSs.SubDomainMPS — Type

An MPS with a projector.

Base.:& — Method

a & b represents the intersection of the indices that a and b are projected at.

Base.:+ — Method

Add two PartitionedMPS objects.

If the two projects have the same projectors in the same order, the resulting PartitionedMPS will have the same projectors in the same order. By default, we use directsum algorithm to compute the sum and no truncation is performed.

Base.:< — Method

a < b means that a is projected at a subset of the indices that b is projected at.

Base.:| — Method

a | b represents the union of the indices that a and b are projected at.

If a is projected at inds=1 and b is not projected for the same inds, then a | b is not projected for inds.

Base.getindex — Method

Indexing for PartitionedMPS. This is deprecated and will be removed in the future.

Base.isdisjoint — Method

Return if projectors are not overlapping

Base.keys — Method

Return the keys, i.e., projectors of the PartitionedMPS.

Base.length — Method

Get the number of sites in the PartitionedMPS

Base.values — Method

Return the values, i.e., SubDomainMPS of the PartitionedMPS.

LinearAlgebra.norm — Method

Return the norm of the PartitionedMPS.

PartitionedMPSs._add_patching — Method

Add multiple SubDomainMPS objects on the same projector.

If the bond dimension of the result reaches maxdim, perform patching recursively to reduce the bond dimension.

PartitionedMPSs._next_projindex — Method

Return the next index to be projected.

PartitionedMPSs.adaptive_patching — Method

Adaptive patching

Do patching recursively to reduce the bond dimension. If the bond dimension of a SubDomainMPS exceeds maxdim, perform patching.

PartitionedMPSs.adaptive_patching — Method

Adaptive patching

Do patching recursively to reduce the bond dimension. If the bond dimension of a SubDomainMPS exceeds maxdim, perform patching.

PartitionedMPSs.adaptivecontract — Function

Perform contruction of two PartitionedMPS objects.

The SubDomainMPS objects of each PartitionedMPS do not overlap with each other. This makes the algorithm much simpler

PartitionedMPSs.add_patching — Method

Add multiple PartitionedMPS objects.

PartitionedMPSs.contract — Method

Contract two Blocked MPS objects.

At each site, the objects must share at least one site index.

PartitionedMPSs.extractdiagonal — Method

Extract diagonal of the PartitionedMPS for s, s', ... for a given site index s, where s must have a prime level of 0.

PartitionedMPSs.makesitediagonal — Method

Make the PartitionedMPS diagonal for a given site index s by introducing a dummy index s'.

PartitionedMPSs.projcontract — Method

Project two SubDomainMPS objects to proj before contracting them. The results are summed.

PartitionedMPSs.projcontract — Method

Project two SubDomainMPS objects to proj before contracting them.

PartitionedMPSs.project — Method

Project the LazyContraction object to prj before evaluating it.

This may result in projecting the external indices of a and b.

PartitionedMPSs.rearrange_siteinds — Method

Rearrange the site indices of the PartitionedMPS according to the given order. If nessecary, tensors are fused or split to match the new order.

PartitionedMPSs.siteindices — Method

Return the site indices of the PartitionedMPS. The site indices are returned as a vector of sets, where each set corresponds to the site indices at each site.